# dashboard/openapi.json
{
  "components": {
    "schemas": {
...
      "CalendarDayOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "density": {
            "title": "Density",
            "type": "integer"
          },
          "problems": {
            "title": "Problems",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "count",
          "problems",
          "density"
        ],
        "title": "CalendarDayOut",
        "type": "object"
      },
      "CalendarOut": {
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/CalendarDayOut"
            },
            "title": "Days",
            "type": "array"
          },
          "month": {
            "title": "Month",
            "type": "string"
          },
          "today": {
            "title": "Today",
            "type": "string"
          }
        },
        "required": [
          "month",
          "today",
          "days"
        ],
        "title": "CalendarOut",
        "type": "object"
      },
...
      "DayOut": {
        "properties": {
          "available_types": {
            "items": {
              "type": "string"
            },
            "title": "Available Types",
            "type": "array"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "hours": {
            "items": {
              "$ref": "#/components/schemas/HourOut"
            },
            "title": "Hours",
            "type": "array"
          },
          "today": {
            "title": "Today",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_bytes": {
            "title": "Total Bytes",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "today",
          "total",
          "total_bytes",
          "hours",
          "available_types"
        ],
        "title": "DayOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HourOut": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "hour": {
            "title": "Hour",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "hour",
          "total",
          "bytes"
        ],
        "title": "HourOut",
        "type": "object"
      },
...
      "Page_RecordingOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecordingOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          }
        },
        "required": [
          "items",
          "next"
        ],
        "title": "Page[RecordingOut]",
        "type": "object"
      },
      "RecordingOut": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "ciphertext_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ciphertext Size"
          },
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_error_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Class"
          },
          "last_error_class_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Class Text"
          },
          "last_error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Detail"
          },
          "local_datetime": {
            "title": "Local Datetime",
            "type": "string"
          },
          "local_time": {
            "title": "Local Time",
            "type": "string"
          },
          "next_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Attempt At"
          },
          "plaintext_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plaintext Sha256"
          },
          "plaintext_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plaintext Size"
          },
          "remote_name": {
            "title": "Remote Name",
            "type": "string"
          },
          "remote_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Size"
          },
          "start_utc": {
            "format": "date-time",
            "title": "Start Utc",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "types": {
            "items": {
              "type": "string"
            },
            "title": "Types",
            "type": "array"
          }
        },
        "required": [
          "id",
          "device_id",
          "remote_name",
          "start_utc",
          "local_time",
          "local_datetime",
          "duration_s",
          "types",
          "state",
          "plaintext_size",
          "ciphertext_size",
          "remote_size",
          "plaintext_sha256",
          "archived_at",
          "attempts",
          "next_attempt_at",
          "last_error",
          "last_error_class",
          "last_error_class_text",
          "last_error_detail"
        ],
        "title": "RecordingOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyOut": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "recording": {
            "$ref": "#/components/schemas/RecordingOut"
          }
        },
        "required": [
          "ok",
          "recording"
        ],
        "title": "VerifyOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ReoVault",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/calendar": {
      "get": {
        "operationId": "get_calendar_api_v1_calendar_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "title": "Month",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Calendar",
        "tags": [
          "footage"
        ]
      }
    },
    "/api/v1/day": {
      "get": {
        "description": "`date_` defaults to the camera's own local today, so the dashboard\ncan open Footage without first knowing the camera's timezone.",
        "operationId": "get_day_api_v1_day_get",
        "parameters": [
          {
            "in": "query",
            "name": "date_",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date "
            }
          },
          {
            "in": "query",
            "name": "types",
            "required": false,
            "schema": {
              "default": "",
              "title": "Types",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Day",
        "tags": [
          "footage"
        ]
      }
    },
    "/api/v1/day/hour": {
      "get": {
        "operationId": "get_day_hour_api_v1_day_hour_get",
        "parameters": [
          {
            "in": "query",
            "name": "date_",
            "required": true,
            "schema": {
              "title": "Date ",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hour",
            "required": true,
            "schema": {
              "maximum": 23,
              "minimum": 0,
              "title": "Hour",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "types",
            "required": false,
            "schema": {
              "default": "",
              "title": "Types",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": "",
              "title": "After",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RecordingOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Day Hour",
        "tags": [
          "footage"
        ]
      }
    },
...
    "/api/v1/recordings/{recording_id}": {
      "get": {
        "operationId": "get_recording_api_v1_recordings__recording_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "title": "Recording Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recording",
        "tags": [
          "footage"
        ]
      }
    },
    "/api/v1/recordings/{recording_id}/retry": {
      "post": {
        "description": "Makes a failed recording eligible for the next run's pickup; it is\nnot re-downloaded on the spot (see Repository.retry_recording).",
        "operationId": "retry_recording_api_v1_recordings__recording_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "title": "Recording Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Recording",
        "tags": [
          "footage"
        ]
      }
    },
    "/api/v1/recordings/{recording_id}/stream": {
      "get": {
        "operationId": "stream_recording_api_v1_recordings__recording_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "title": "Recording Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "download",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Download",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Recording",
        "tags": [
          "footage"
        ]
      }
    },
    "/api/v1/recordings/{recording_id}/verify": {
      "post": {
        "operationId": "verify_recording_api_v1_recordings__recording_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "title": "Recording Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Recording",
        "tags": [
          "footage"
        ]
      }
    },
...
  }
}